fix(video-layout): functions imports.
