+++ react/features/video-layout/functions.native.js
@@ -1,0 +1,1 @@
+export * from './functions.any';
